+++ lib/src/home/view/screens/home_screen.dart
@@ -20,7 +20,7 @@ class HomeScreen extends WebPageBase {
   }
 
   @override
-  Widget? child(BuildContext context) {
+  Widget? child(BuildContext context, [WebPage? widget]) {
     // TODO: implement child
     throw UnimplementedError();
   }
@@ -207,7 +207,7 @@ class HomeScreenController extends WebPageBaseController {
   }
 
   @override
-  Widget? child(BuildContext context) {
+  Widget? child(BuildContext context, [WebPage? widget]) {
     // TODO: implement child
     throw UnimplementedError();
   }